refactor: move InlineSpriteSheet
