+++ src/components/AppComponents/Logo/Logo.tsx
@@ -1,7 +1,7 @@
 import { FC } from "react";
 import styled from "styled-components";
 
-import { InlineSpriteSvg } from "@/components/InlineSpriteSheet";
+import { InlineSpriteSvg } from "@/components/GenericPagesComponents/InlineSpriteSheet";
 import ScreenReaderOnly from "@/components/SharedComponents/ScreenReaderOnly";
 
 type LogoProps = {
