+++ src/components/GenericPagesComponents/InlineSpriteSheet/InlineSpriteSheet.tsx
@@ -6,7 +6,7 @@
 import { FC } from "react";
 
 // eslint-disable-next-line import/no-unresolved
-import InlineSpriteSheetSvg from "../../image-data/generated/inline-sprite.svg";
+import InlineSpriteSheetSvg from "@/image-data/generated/inline-sprite.svg";
 
 const InlineSpriteSheet: FC = () => {
   return <InlineSpriteSheetSvg />;
